[WIP] move models to have a mtching naming scheme
