+++ app/src/main/java/com/github/mrpaulblack/personalrecipes/data/models/IngredientsModel.kt
@@ -1,0 +1,12 @@
+package com.github.mrpaulblack.personalrecipes.data.models
+
+data class IngredientsModel (
+    val quantity: Double = 0.0,
+    val weight: Double = 0.0,
+    val text: String = "",
+    val image: String = "",
+    val measure: String = "",
+    val foodId: String = "",
+    val foodCategory: String = "",
+    val food: String = ""
+)
