+++ app/src/main/java/com/github/mrpaulblack/personalrecipes/data/models/RecepiesOverviewListModel.kt
@@ -1,0 +1,5 @@
+package com.github.mrpaulblack.personalrecipes.data.models
+
+class RecepiesOverviewListModel {
+    var recepies: MutableList<RecipesDetailModel> = mutableListOf<RecipesDetailModel>()
+}
