+++ app/src/main/java/com/github/mrpaulblack/personalrecipes/data/models/RecipeModel.kt
@@ -1,0 +1,6 @@
+package com.github.mrpaulblack.personalrecipes.data.models
+
+data class RecipeModel(
+    val title: String,
+    val text: String
+)
